Fix log file name and .env file
Windows does not like colons in the filename. It also gets confused when you copy the **entire file path** into the file name.

As for .env, the file just wasn't in my repo and the is_dev function didn't catch it. I changed the .gitignore to fix this, but we should address why it was in the gitignore in the first place